pyproject.toml

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voice-assistant"
version = "0.1.0"
description = "A low-latency voice assistant that listens for commands and responds"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Vignesh Srinivasan", email = "vignesh.mssrinivasan@example.com" }
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "fastrtc",  # Note: You may need to specify the exact version
    "fastrtc[stt,vad,tts,stopword]",
    "gradio>=3.0.0",
    "python-dotenv>=0.19.0",
    "pydantic>=1.8.0",
    "llama-cpp-python",
    "fastrtc-whisper-cpp",
    "python-multipart"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=4.0.0",
    "mypy>=0.900",
]

[project.scripts]
assistant = "src.main:main"

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q" 